- Fixed type in configuration.cpp and configation.h from unsigned int to size_t. - Removed check for xrc-library, as this may not exist in a seperate file.

+++ trunk/mapgenerator/src/util/configuration.cpp
@@ -187,7 +187,7 @@ namespace mapgeneration_util
 
     
 	void
-	Configuration::characters(const unsigned char *text, unsigned int len)
+	Configuration::characters(const unsigned char *text, size_t len)
 	{
 		std::string value(reinterpret_cast<const char*>(text), len);
 		_parse_stack.back()._value = value;
